Intentar definir las clases de clientes y TipoClientes en app usuarios; Eliminar el app de accounts y archivos relacionados
