+++ .env.dev
@@ -9,3 +9,4 @@ SQL_DATABASE=postgres # dev database
 SQL_HOST=db
 SQL_PORT=5432
 DATABASE=postgres
+DJANGO_SETTINGS_MODULE=global_exchange_django.settings
