--- app/apps/accounts/templates/accounts/login.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <title>Login</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-100">
-    <div class="container mx-auto mt-10 max-w-md bg-white p-6 rounded shadow">
-        <h2 class="text-2xl font-bold mb-4 text-center">Login</h2>
-        <form method="POST" class="space-y-4">
-            {% csrf_token %}
-            <input type="hidden" name="next" value="{% url 'home' %}">
-            <div>
-                <label for="username" class="block font-medium">Username</label>
-                <input type="text" name="username" id="username" class="w-full border rounded px-3 py-2" required>
-            </div>
-            <div>
-                <label for="password" class="block font-medium">Password</label>
-                <input type="password" name="password" id="password" class="w-full border rounded px-3 py-2" required>
-            </div>
-            <button type="submit" class="w-full bg-blue-500 text-white py-2 rounded hover:bg-blue-700">Login</button>
-        </form>
-        <p class="mt-4 text-center">¿No tienes cuenta? <a href="{% url 'register' %}" class="text-blue-500">Register</a></p>
-    </div>
-</body>
-</html>
