--- app/apps/accounts/urls.py
@@ -1,15 +1,0 @@
-from django.contrib.auth import views as auth_views
-from django.urls import path
-
-from . import views
-
-urlpatterns = [
-    path("", views.home, name="home"),  # TODO: Modulo de presentacion?
-    path("register/", views.register, name="register"),
-    path(
-        "login/",
-        auth_views.LoginView.as_view(template_name="accounts/login.html", redirect_authenticated_user=True),
-        name="login",
-    ),
-    path("logout/", auth_views.LogoutView.as_view(next_page="home"), name="logout"),
-]
