+++ app/apps/panel_admin/urls.py
@@ -13,10 +13,10 @@ urlpatterns = [
     path("usuarios/crear/", views.usuario_create, name="usuario_crear"),
     path("usuarios/<int:pk>/editar/", views.usuario_edit, name="usuario_editar"),
     path("usuarios/<int:pk>/eliminar/", views.usuario_delete, name="usuario_eliminar"),
-    path("roles/", views.rol_list, name="rol_listar"),
-    path("roles/crear/", views.rol_create, name="rol_crear"),
-    path("roles/<int:pk>/editar/", views.rol_edit, name="rol_editar"),
-    path("roles/<int:pk>/eliminar/", views.rol_delete, name="rol_eliminar"),
+    # path("roles/", views.rol_list, name="rol_listar"),
+    # path("roles/crear/", views.rol_create, name="rol_crear"),
+    # path("roles/<int:pk>/editar/", views.rol_edit, name="rol_editar"),
+    # path("roles/<int:pk>/eliminar/", views.rol_delete, name="rol_eliminar"),
     path("clientes/", views.cliente_list, name="cliente_listar"),
     path("clientes/crear/", views.cliente_create, name="cliente_crear"),
     path("clientes/<int:pk>/editar/", views.cliente_edit, name="cliente_editar"),
