+++ app/apps/usuarios/models.py
@@ -1,4 +1,9 @@
-from django.contrib.auth.models import AbstractBaseUser, BaseUserManager, PermissionsMixin
+from django.conf import settings
+from django.contrib.auth.models import (
+    AbstractBaseUser,
+    BaseUserManager,
+    PermissionsMixin,
+)
 from django.db import models
 
 
@@ -62,6 +67,36 @@ class Rol(models.Model):
     nombre = models.CharField(max_length=30, unique=True)
     descripcion = models.TextField(blank=True)
     permisos = models.ManyToManyField(Permiso, blank=True)
+
+    def __str__(self):
+        return self.nombre
+
+
+class TipoCliente(models.Model):
+    """Modelo que representa el tipo de cliente."""
+
+    nombre = models.CharField(max_length=50, unique=True)
+
+    def __str__(self):
+        return self.nombre
+
+
+class Cliente(models.Model):
+    """Modelo que representa a un cliente."""
+
+    ruc = models.CharField(max_length=11, unique=True)
+    nombre = models.CharField(max_length=100)
+    email = models.EmailField(unique=True)
+    telefono = models.CharField(max_length=15, blank=True)
+    direccion = models.CharField(max_length=255, blank=True)
+
+    # referencia a TipoCliente
+    tipo_cliente = models.ForeignKey(
+        TipoCliente, on_delete=models.SET_NULL, null=True, blank=True
+    )
+
+    # referencia dinámica al modelo de usuario
+    usuarios = models.ManyToManyField(settings.AUTH_USER_MODEL, related_name="clientes")
 
     def __str__(self):
         return self.nombre
